+++ shared-libs/crates/start-core/src/util/mod.rs
@@ -38,7 +38,6 @@ pub mod collections;
 pub mod cpupower;
 pub mod crypto;
 pub mod data_url;
-pub mod direct_io;
 pub mod future;
 pub mod http_reader;
 pub mod io;
@@ -56,6 +55,7 @@ pub mod serde;
 pub mod sync;
 pub mod tui;
 pub mod version;
+pub mod writeback;
 
 pub use clap::FromStrParser;
 pub use data_url::DataUrl;